+++ R/updateDesign.R
@@ -1,12 +1,12 @@
 #' Update design elements
 #' 
-#' A function that will update the object returned from \code{\link{findNextItem}}.
+#' A function that will update the object returned from \code{\link{findNextItem}}. This can 
+#' be used to run the CAT session manually in a set-by-step manner. 
 #' 
 #' @param x an object of class 'mirtCAT_design' returned from the \code{\link{mirtCAT}} function
 #'   when passing \code{design_elements = TRUE}
-#' @param items a numeric vector indicating which items to select
-#' @param responses a numeric vector indicating the responses the the selected items
-#' @param Theta (optional) vector indicating the value of Theta/latent traits to be set
+#' @param new_item a numeric vector indicating which items to select
+#' @param new_response a numeric vector indicating the responses the the selected items
 #' 
 #' @seealso \code{\link{mirtCAT}}, \code{\link{findNextItem}}
 #' @export updateDesign
@@ -23,33 +23,70 @@
 #' @return returns an object of class 'mirtCAT_design' with updated elements.
 #' @examples
 #' \dontrun{
+#' 
+#' set.seed(1)
+#' nitems <- 100
+#' itemnames <- paste0('Item.', 1:nitems)
+#' a <- matrix(rlnorm(nitems, .2, .3))
+#' d <- matrix(rnorm(nitems))
+#' dat <- simdata(a, d, 500, itemtype = 'dich')
+#' colnames(dat) <- itemnames
+#' mod <- mirt(dat, 1, verbose = FALSE)
+#' 
 #' # test defined in mirtCAT help file, first example
-#' CATdesign <- mirtCAT(df, mod, criteria = 'MI', design_elements = TRUE)
+#' CATdesign <- mirtCAT(mo = mod, criteria = 'MI', design_elements = TRUE,
+#'   start_item = 2)
 #' 
-#' # returns number 1 in this case, since that's the starting item
+#' # returns 2 in this case, since that was the starting item
 #' findNextItem(CATdesign) 
 #' 
-#' # determine next item if item 1 and item 10 were answered correctly, and Theta = 0.5
-#' CATdesign <- updateDesign(CATdesign, items = c(1, 10), responses = c(1, 1), Theta = 0.5)
-#' findNextItem(CATdesign) 
+#' # first iteration, no answered items
+#' CATdesign$person$items_answered
 #' 
-#' # alternatively, update the Theta using the Update.thetas definition in design
+#' # update when next item is item 2 and answered correctly
+#' CATdesign <- updateDesign(CATdesign, new_item = 2, new_response = 1)
+#' CATdesign$person$items_answered  # item 2 answered first
+#' CATdesign$person$responses       # in item 2 element response was = 1 
+#' 
+#' # update internal Theta estimate using Update.thetas definition
 #' CATdesign$design@Update.thetas(CATdesign$design, CATdesign$person, CATdesign$test) 
+#' CATdesign$person$thetas
+#' findNextItem(CATdesign) # find next item
+#' 
+#' # determine next item if item 70 were also answered correctly next
+#' CATdesign <- updateDesign(CATdesign, new_item = 70, new_response = 1)
+#' CATdesign$person$items_answered  
+#' CATdesign$person$responses       
+#' 
+#' # update the Theta using the Update.thetas definition in design
+#' CATdesign$design@Update.thetas(CATdesign$design, CATdesign$person, CATdesign$test) 
+#' CATdesign$person$thetas # current estimate
+#' CATdesign$person$thetas_history
+#' CATdesign$person$thetas_SE_history
+#' findNextItem(CATdesign) # find next item
+#' 
+#' # continue on, now with item 95 added next (answered incorrectly)
+#' CATdesign <- updateDesign(CATdesign, new_item = 95, new_response = 0)
+#' CATdesign$design@Update.thetas(CATdesign$design, CATdesign$person, CATdesign$test) 
+#' CATdesign$person$thetas
+#' CATdesign$person$thetas_history
+#' CATdesign$person$thetas_SE_history
 #' findNextItem(CATdesign)
+#' 
 #' }
 #' 
-updateDesign <- function(x, items, responses, Theta=NULL){
-    if(missing(x) || missing(items) || missing(responses))
+updateDesign <- function(x, new_item, new_response){
+    if(missing(x) || missing(new_item) || missing(new_response))
         stop('require inputs have not been supplied', call.=FALSE)
-    if(!is.null(Theta)){
-        Theta <- matrix(Theta, nrow = 1L)
-        x$person$thetas <- Theta
-    }
-    if(any(items > length(x$person$responses)))
+    if(length(new_item) > 1L || length(new_response) > 1L)
+        stop("new_item and new_response must be length 1 integers")
+    if(new_item > length(x$person$responses))
         stop('Items locations are larger than the length of the test.', call.=FALSE)
-    x$person$responses[items] <- as.integer(responses)
-    x$person$raw_responses[items] <- as.character(x$person$responses[items] + 1L)
+    if(any(x$person$items_answered %in% new_item))
+        stop('The new_item selected has already been administered', call.=FALSE)
+    x$person$responses[new_item] <- as.integer(new_response)
+    x$person$raw_responses[new_item] <- as.character(x$person$responses[new_item] + 1L)
     pick <- min(which(is.na(x$person$items_answered)))
-    x$person$items_answered[pick:(length(responses)+pick-1L)] <- as.integer(items)
+    x$person$items_answered[pick] <- as.integer(new_item)
     return(x)
 }